--- presentation/src/main/java/com/woowahan/accountbook/ui/screens/setting/mode/SettingMode.kt
@@ -1,6 +1,0 @@
-package com.woowahan.accountbook.ui.screens.setting.mode
-
-enum class SettingMode {
-    Create(),
-    Modify()
-}
