+++ presentation/src/main/java/com/woowahan/accountbook/ui/type/SettingMode.kt
@@ -1,0 +1,6 @@
+package com.woowahan.accountbook.ui.type
+
+enum class SettingMode {
+    Create(),
+    Modify()
+}
